requirements.txt
amqp>=5.3.1
asgiref>=3.11.1
attrs>=26.1.0
billiard>=4.2.4
celery>=5.6.3
celery-types>=0.26.0
certifi>=2026.2.25
click>=8.3.2
click-didyoumean>=0.3.1
click-plugins>=1.1.1.2
click-repl>=0.3.0
cron-descriptor>=1.4.5
dj-database-url>=3.1.2
Django>=6.0.4
django-celery-beat>=2.9.0
django-cors-headers>=4.9.0
django-timezone-field>=7.2.1
djangorestframework>=3.17.1
djangorestframework_simplejwt>=5.5.1
drf-spectacular>=0.29.0
gunicorn>=25.3.0
inflection>=0.5.1
jsonschema>=4.26.0
jsonschema-specifications>=2025.9.1
kombu>=5.6.2
packaging>=26.0
prompt_toolkit>=3.0.52
psycopg2-binary>=2.9.11
PyJWT>=2.12.1
python-crontab>=3.3.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.2
PyYAML>=6.0.3
redis>=7.4.0
referencing>=0.37.0
rpds-py>=0.30.0
sentry-sdk>=2.57.0
six>=1.17.0
sqlparse>=0.5.5
typing_extensions>=4.15.0
tzdata>=2026.1
tzlocal>=5.3.1
uritemplate>=4.2.0
urllib3>=2.6.3
vine>=5.1.0
wcwidth>=0.6.0
whitenoise>=6.12.0
